RDCC-1011 RouteProperty class moved to route and fixed IMapper
